Pass user command text to the Command object
Previously the Command objects did not know what command text was entered
by the user.  That mapping is done in entry points and was thrown away in
App.run_subcommand().

This adds cmd_name as an optional keyword arg to Command.__init__() and
saves it as an attribute.  App.run_subcommand() passes in the original
command text to cmd_factory() and all is well.

Closes-bug: #1411095
Change-Id: I36c42b0447652ed3bd71ad6ae0d70db7d27abbad