app,command: disallow abbrev by default
By default, abbrev support is enabled in argparse, which can mangle
cliff arguments as shown in the test since all the arguments from all
the commands are not registered at once in the parser, but are parsed in
multiple passes.

It's a nasty behaviour that by default can break application using cliff
without having them noticing. Since I don't think anyone is really
relying on the abbrev behaviour for real, let's just disable it
altogether.

Depends-On: I903ce9564a83d3ee69f4efeb726d3c2d3ff69bbb
Change-Id: I77127e364e938418de36b09c0b80644b77d62336