Add CLIFF_MAX_TERM_WIDTH envvar to complement --max-width
Some cliff clients tailors their output according to characteristics
of the terminal they outputs text to. They detect the width of the
terminal in order to pretty print text properly cut to line length.

There are situations where the output is not in a "real" terminal
but in some kind of emulation or redirected layer for which this
line length interpretation is not ideal. For example in a bash
jupyter-notebook.

There already exist the --max-width CLI parameter to allow that
to be customized, but there are cases where you don't want to
use an additional parameter. So add an environment variable
that has the same meaning, to help avoiding the use of
--max-width.

Without the parameter nor the envvar, the behaviour should stay
the same as before the patch, i.e. the output is cut to terminal
line length. The envvar is then used to give the parameter's
default value, so the parameter will stay being the final one
deciding of the behaviour when used.

Add mock'ed testing for the new CLIFF_MAX_TERM_WIDTH envvar.
Add mock'ed tests to ensure parameter vs envvar priority.
Add testing for actual parameter parsing & priority vs envvar.²

This has been discussed in:
https://bugs.launchpad.net/python-openstackclient/+bug/1577368
https://bugs.launchpad.net/python-cliff/+bug/1485847/comments/7

Closes-Bug: 1577368
Related-Bug: 1485847

Change-Id: I0ab988ea0e7087960a9bfb2353f46918f0333a53
Signed-off-by: Vincent Legoll <vincent.legoll@idgrilles.fr>