add formattable columns concept
Some items that need to be treated as individual values are actually
complex data structures. Different formatters need to render those
values in different ways. For human-readable formatters like the table,
we may want to strip punctuation from data structures, for example,
while JSON formatters may need the values to be machine-readable and it
may not be possible to add extra punctuation or whitespace.

This patch adds the ability to all built-in formatters to have values
returned from commands be FormattableColumn instances, which can then be
interrogated for the two different forms of the data (human or machine
readable).

Change-Id: Ic84b05ae720188a63a11151e012997416544ac34
Signed-off-by: Doug Hellmann <doug@doughellmann.com>