Add Constraints support
Adding constraints support to libraries is slightly more complex than
services as the libraries themselves are listed in upper-constraints.txt
which leads to errors that you can't install a specific version and a
constrained version.

This change adds constraints support by also adding a helper script to
edit the constraints to remove cliff and at the same time pull any required
libraries from git.

Also the egg name fro python-neutronclient is wrong resulting in
warnings in the logs[1] so correct that wheil re-writing the deps in
tox.ini

[1] http://logs.openstack.org/42/412742/1/check/gate-cliff-tox-neutronclient-tip-ubuntu-xenial/e1ce5c0/console.html#_2016-12-20_04_40_50_638986

Change-Id: Ifa5cc8446cfbfec0c8b48b293feb76ce66a34d92