sphinxext: Generate better usage examples
At the moment, the 'autoprogram-cliff' directive uses the default usage
formatter in its output. This is functional but no more. Take, for
example, the following output, generated for the python-openstackclient
project:

  openstack server migrate [--live <hostname>]
                           [--shared-migration | --block-migration]
                           [--disk-overcommit | --no-disk-overcommit] [--wait]
                           <server>

What we actually want is something like the below, which is not only
more legible but also avoids issues like optional argument flags and
the corresponding values being on separate lines:

  openstack server migrate
      [--live <hostname>]
      [--shared-migration | --block-migration]
      [--disk-overcommit | --no-disk-overcommit]
      [--wait]
      <server>

Do this by manually generating our own usage samples, harnessing a
little of the internal argparse infrastructure.

Change-Id: If4dff4991562da9892f1c06f854143b71111007a