sphinxext: Add 'application' option to the autoprogram directive
In some cases, a single documentation build wants to handle multiple
cliff-based CLI. One possible example is openstack-manuals CLI ref.
In this case, the global setting 'autoprogram_cliff_application'
does not work. It is nice if we can specify an application name
per autoprogram-cliff directive.
This commit introduces a new option 'application' to
autoprogram-cliff directive to allow this.

Change-Id: Ie2d01920dc04f2a92031a641d809f7da8af8b395