make smart help formatter test deterministic
Set up the formatter to always use a width=80 so that the terminal width
of the developer's system does not cause the test to fail. Trying to
mock os.environ failed, but there is an arg to HelpFormatter to set the
width explicitly. Unfortunately, there is no way to do that through the
parser, so we have to replace the parser's formatter_class attribute
with a partial() that passes width to the original class.

Change-Id: I2ab035b5b188f1a028e1776ba9f09c71dc316c11
Signed-off-by: Doug Hellmann <doug@doughellmann.com>