sphinxext: Support cliff application
This commit adds a feature to generate a reference of global
optoions of a cliff application to autoprogram-cliff directive.
If a class path of a cliff application is specified as the argument of
autoprogram-cliff directive, it will be interpreted as a cliff application
and global options of the specified application are rendered.

Change-Id: I20e46521a137ca721fae28f10c5cf75d26069e45