Alias exit to cmd2's quit command to exit interactive shell
`exit` is a user friendly and much more intuitive way to quit the
interactive shell. It is consistent with the Python interpreter,
IPython and even OSC which provides an exit command. Applications
created using Cliff do not need to implement it. It has the same
behavior as `quit` command as it is just another name for cmd2's
`do_quit` method.

Change-Id: I19ca9d2303379838e5fbd81e9d386832727ea618