add argparse conflict handler "ignore"
Update our version of ArgumentParser with a conflict resolution
handler called "ignore" to ignore options from commands if they would
conflict with options already registered. An error is reported if the
action associated with the option would not be registered at all
because all of its names conflict. A warning is reported for each
option string that is being ignored.

Change-Id: I99c62d5772017333136527f7f509c776623641a1
Signed-off-by: Doug Hellmann <doug@doughellmann.com>