Capturing argparse errors due to problem with cmd2
The Bifrost team got an errors on their CLI that also affects OSC. When
in interactive mode, if argparse fails to parse the input due to, say,
a missing parameter, argparse by default thows a SystemExit(2), but
cmd2 doesn't like it because that could've been a signal to stop the
CLI, so it breaks the interactive session. This fix aims to bypass that
and keep the CLI running so we don't have to start it at every parameter
we forget to type in.

Change-Id: I0e2006a9625e2f8dbdbc0e5921acfb3853a06ee9
Story: 2008071
Task: 40782