Remove lower-constraints
This is not part of the PTI and is currently broken. While discussions
are ongoing about removing it from every project, there's a definite
lean towards doing so. Let's do just that. We can re-add in the future
if necessary.

While we're here, we fix some indentation in 'tox.ini'.

Change-Id: Ib4784d9da96d05a54acdfbb3744af0cb053c0c6c
Signed-off-by: Stephen Finucane <sfinucan@redhat.com>