Add conflict_handler parameter as attribut in Command class
Adding conflict_handler as attribut in the Command class in order to be
able to take control of this parameter and change to different behavior
that argparse is handling: error / resolve / ignore.

Callers will be able to override it and get a proper Parser object.

Change-Id: I327ece99a04bc8b2ebfa554dea643b1f2a456336
(cherry picked from commit 452fff3aadb01c41355a9254c6f21b76b5ed1609)