Removing modindex link from docs
The global module documentation is no longer being generated
and the link points to a non-existent document.

The purpose of global module documentation is now fulfilled by
the automatically generated class, and function level reference
documents which are linked properly.

As such there is no need to keep the link around, or to
reestablish module level documentation.

Signed-off-by: Jiri Podivin <jpodivin@redhat.com>
Change-Id: I05e5144a36f33aa5feb996964d1a098b1716cf6a