Use upper-constraint in doc generation
Currently doc jobs are broken because the latest sphinx 7.2.0 is not
compatible with the latest release of openstackdocstheme.

This updates the docs/releasenotes target to use the global upper
constraints to avoid that issue.

Also, the following flake8 error is fixed by this change.

cliff/tests/test_formatters_table.py:53:10: E721 do not compare types,
for exact checks use `is` / `is not`, for instance checks use
`isinstance()`

Change-Id: Ia923ba1d0e2cf3327317e142cfa38470d9f053d2