Autofit table output if stdout is a tty
nobody enjoys reading mangled tables in the terminal.

Be smart and auto-enable "fit width" if the output is a tty,
while still printing full lines when output is redirected (e.g. pipe)
so that e.g. grep and friends continue to work as before.

Unfortunately, tty detection is not reliable enough on Windows,
so this feature is not enabled there.

Change-Id: I46137d16ea3054a43de4a9f76477bda59ebacf10