typing: Add typing to cliff.command
We also type all the in-tree commands, since it tests our signatures for
the base classes.

The only potentially controversial change we have here is that we no
longer set 'self.app_args.cmd' in 'HelpCommand'. I'm not entirely sure
why we were doing this: it seems to have been introduced as part of the
fix for bug #1421585 [1], but removing it causes no harm and I don't see
it doing anything useful.

[1] https://bugs.launchpad.net/python-cliff/+bug/1421585

Change-Id: I7b889a775f5b61b28b4b22c866b7301393206e7d
Signed-off-by: Stephen Finucane <sfinucan@redhat.com>