hooks: Update type hints to reflect reality
Despite the docstrings (and now type hints) indicating that 'get_epilog'
and 'get_parser' methods should return a 'str' and an
'argparse.ArgumentParser' instance, respectively, there are users in the
wild (like Kayobe's 'HookDispatcher' returning None for both). This
works because there is logic within cliff itself to exclude 'None'
results from the former, while the return result from the latter is
never used.

Update the docs and type hints to reflect reality.

Change-Id: If8c7deee307b61ee709b8bbefae7e93a9f651e4c
Signed-off-by: Stephen Finucane <sfinucan@redhat.com>