Remove pkg_resources remnants
stevedore switched its internal implementation to use importlib.metadata
instead of pkg_resources in v3.0.0 [1]. The EntryPoint class provided by
'importlib.metadata' only provides a 'load' method that takes no
arguments [2]. There is therefore no reason to check for a 'resolve'
method, nor to check for a 'require' argument that is not needed for
'importlib.metadata'. Bump the stevedore minimum version to v3.0.0 and
drop this logic.

[1] https://opendev.org/openstack/stevedore/commit/d5297167e08468c75d2477f15004df61cf98e57e
[2] https://docs.python.org/3/library/importlib.metadata.html#importlib.metadata.EntryPoint

Change-Id: I09ae40afe9f26659d44e26a16d10f1ff4eb57d1d
Signed-off-by: Stephen Finucane <sfinucan@redhat.com>